requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
dj-rest-auth==7.0.1
Django==5.2.9
django-allauth==65.13.1
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
outscraper==5.10.0
psycopg2==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
sqlparse==0.5.4
urllib3==2.5.0
whitenoise==6.11.0
